Fix XCTest UI testing framework - working screenshot system
- Add accessibility identifiers to UI elements for reliable test targeting
- Fix sleep() function calls to use UInt32 parameter type
- Add GENERATE_INFOPLIST_FILE to UI test target configuration
- Fix XCTAttachment.add() method call in ScreenshotManager
- Successfully generating test screenshots in test_results/ directory
- Single test (AppLaunchTests/testAppLaunches) now passes and captures UI state

+++ UITests/CustomTests/ExampleCustomTest.swift
@@ -53,13 +53,13 @@ class ExampleCustomTest: XCTestCase {
             let urlField = app.textFields["URL"]
             urlField.tap()
             urlField.typeText("https://")
-            sleep(1)
+            sleep(UInt32(1))
             urlField.typeText("httpbin.org/get")
         }
         
         framework.runCustomAction(name: "manual_go_button") {
             app.buttons["Go"].tap()
-            sleep(3)
+            sleep(UInt32(3))
         }
         
         framework.openDevTools(screenshotName: "custom_devtools")
